consolidated fixes into ceramic-cli-utils
- as per comments suggestion;
